Both browser and localgui monitors now working cleaner
Refactored some of the code to make it more consistent across monitor types. Still need to cleanup the design of the monitoring to make it more flexible, and break it loose from the main code. Might still want to consider running the monitor parts on the main thread, and use other threads to run other logic.