Removed lots of old code/features ...
To hopefully make it easier to clean up the code for the new version. Removed concepts of "Shot, Div, Alias and Shotlist". I will try to reproduce with namespace elements. This allowed me to remove the peekAhead logic, and all the handling code. Always pass the args along to avoid losing context.