Added ability to generate custom head markdown in smdparse & ismd
This got broken when the way builtins were imported at startup. The new way fits the new model of overriding one or more of the document defaults on the command line. Had to add ability to silence the thread id printing in the console output, otherwise it breaks between runs. Added a hack for comparing the generated output from smdparse, which will require a refactor later when more tests are added to smdparse.