Added a new code builtin: wrap_stack
This allows you to use the wrap_stack independently of the formatter. In other words, you can emit the start or end portions on demand, or you can print them out to see what they are. This was implemented using the thread-safe TLS logic, which will pave the way for eliminating the remaining globals in smd, enabling it to be fully thread-safe.