Add ability to print debug messages from utility class
This allows me to insert debug messages during the CodeHelpers, which is very useful. Also, it's done using TLS, so this provides a POC that we can remove the globals from this module in favor of adding them to the TLS. Added couple of new methods for updating attributes in place, including one that uses all strings. Fixed a longstanding bug where lines would not split on \n because it was a literal in the string...