markdown.py updates
Added ability to dump simple to read version of recursive strings if debug is on for 'markdown'. This is helpful on the console to see where you've got an infinite loop.

Added more documentation and commented out code for implementing handling markdown of {{vars}} in the just-in-time dictionary, as well as an improvement in how the regex works with ']' embedded in parameter strings.